Move RejectAll policy to RateLimiter package
